Add `PG::Connection#register_decoder`
This refactors the existing behaviour to have a shared interface (which could
be an abstract module) for registering decoders:

    Decoders.register_decoder(decoder : Decoders::Decoder, oid)
    PG::Connection#register_decoder(decoder : Decoders::Decoder, oid)

While the `Decoders` module maintains responsibility for discovering types, it
adds them to the connection's map via this interface, instead of returning a
new `DecoderMap` for the `PG::Connection` to assign.

This also means the connection's `DecoderMap` is never nil and isn't swapped
out. However, it is mutated. This only happens automatically before the
connection is available for use by the caller but the `#register_decoder`
method could be called later.