Add scram channel binding (scram-sha-265-plus)
This will be preferred when the Postgres server offers it, as it is
strictly better than scram without channel binding. This was a new
feature in Postgres 11. Postgres 10 supports scram, but not channel
binding.

It requires a few extra openssl functions linked, which did not become a
part of Crystal until version 1.1.0, so I've backported that patch for
earlier versions of Crystal.