Drop support for postgres 9.6
It fell out of support from the postgres project itself on Nov 11 2021,
so stop running CI tests for it. https://www.postgresql.org/support/versioning/

The postgres protocol is still unchanged, so it still probably works
fine however.

Now that all supported versions support scram, the CI setup can be
simplified a bit, which is nice.